🔀 [FE] merge: from feature/meal-register-fix,camera/402 to dev

See merge request s12-final/S12P31S206!139

+++ FE/mobile/src/main/java/com/example/diaviseo/network/food/dto/req/FoodWithQuantityRequest.kt
@@ -2,5 +2,5 @@ package com.example.diaviseo.network.food.dto.req
 
 data class FoodWithQuantityRequest(
     val foodId: Int,
-    val quantity: Int
+    val quantity: Float
 )
